refactor(#143): consolidate channel/feishu/, remove feishu/ and teammate/ orphan dirs, rename channel provider stub to plugin.ts
