+++ packages/dreamux/src/agent-runtime/builtin/codex/mcp-config.ts
@@ -2,14 +2,14 @@ import type { AgentRuntimeMcpServer } from '../../types.js';
 import {
   feishuMcpServerDescriptor,
   type FeishuMcpCodexArgsOptions,
-} from '../../../channel/feishu-mcp-surface.js';
+} from '../../../channel/feishu/feishu-mcp-surface.js';
 
 export {
   FEISHU_MCP_SERVER_NAME,
   feishuMcpServerDescriptor,
   type FeishuMcpCodexArgsOptions,
   type FeishuMcpServerDescriptorOptions,
-} from '../../../channel/feishu-mcp-surface.js';
+} from '../../../channel/feishu/feishu-mcp-surface.js';
 
 export function feishuMcpCodexArgs(
   opts: FeishuMcpCodexArgsOptions,
