+++ packages/dreamux/src/dispatcher-service/service.ts
@@ -3,7 +3,7 @@ import type { AgentRuntime } from '../agent-runtime/index.js';
 import type { DreamuxConfig } from '../runtime/config.js';
 import type { DispatcherStore } from '../runtime/dispatcher-store.js';
 import type { DreamuxLogger } from '../runtime/logger.js';
-import type { FeishuBot } from '../feishu/bot.js';
+import type { FeishuBot } from '../channel/feishu/bot.js';
 import type { DispatcherRow } from '../runtime/dispatcher-store.js';
 import type { RestartIntentConsumer } from '../daemon/restart-intent.js';
 import {
